requirements.txt
# Production dependencies
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
alembic>=1.13.0
psycopg2-binary>=2.9.9   # PostgreSQL driver (requires libpq; not needed for tests)
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1,<4.3.0     # passlib __about__ attr removed in 4.3+
python-multipart>=0.0.9
redis>=5.0.0
slowapi>=0.1.9
httpx>=0.27.0
email-validator>=2.2.0
python-dotenv>=1.0.1
aiofiles>=23.0.0
apscheduler>=3.10.0
aiosqlite>=0.20.0        # SQLite async driver (used in tests)

# Test dependencies
pytest>=8.0.0
anyio>=4.4.0
